requirements.txt
flask==3.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
werkzeug==3.1.3
python-dotenv==1.0.1
markdown==3.7
gunicorn==22.0.0
requests==2.32.3
